backend/requirements.txt
aioredis==1.3.1
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.6.0
bleach==5.0.1
boto3==1.24.23
botocore==1.27.23
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.1
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
clayful==2.5.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coolsms-python-sdk==2.0.3
cryptography==37.0.4
daphne==3.0.2
distlib==0.3.5
Django==4.1
django-ckeditor==6.4.2
django-cors-headers==3.10.0
django-debug-toolbar==3.6.0
django-filter==21.1
django-hosts==5.1
django-js-asset==2.0.0
django-revproxy==0.10.0
django-storages==1.12.3
django-summernote==0.8.20.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.0
filelock==3.7.1
firebase-admin==5.3.0
gevent==21.8.0
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
greenlet==1.1.2
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.20.4
hyperlink==21.0.0
iamport-rest-client==0.9.0
identify==2.5.2
idna==3.3
importlib-resources==5.9.0
incremental==21.3.0
inflection==0.5.1
isort==5.10.1
jmespath==1.0.1
jsonschema==4.9.1
kombu==5.2.4
msgpack==1.0.4
mypy-extensions==0.4.3
nodeenv==1.7.0
pathspec==0.9.0
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pre-commit==2.20.0
prompt-toolkit==3.0.30
proto-plus==1.22.1
protobuf==4.21.7
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==3.5.3
requests==2.26.0
rsa==4.9
s3transfer==0.6.0
sentry-sdk==1.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
